backend/requirements/base.txt
Django==1.6
South==0.8.4
dj-database-url==0.2.2
dj-static==0.0.5
django-cors-headers==0.12
djangorestframework==2.3.10
gunicorn==18.0
static==0.4
wsgiref==0.1.2
social.apps.django_app.default

